exp encreace process

+++ Assets/Scripts/Items/ExpItem.cs
@@ -5,13 +5,13 @@ namespace Items
 {
     public class ExpItem : MagnetableItem
     {
-        public float Points { get; private set; }
+        public int Points { get; private set; }
         
         [SerializeField] private float swingAngle = 30f; // Угол покачивания
         [SerializeField] private float duration = 1f;
         private Tween _swingTween;
 
-        public void SetPoints(Sprite sprite, float points)
+        public void SetPoints(Sprite sprite, int points)
         {
             _renderer.sprite = sprite;
             Points = points;
